Completed Add, Search, Delete function. Styles are changed as well.

+++ lib/src/hero_tour/hero_tour_list/hero_tour_list_component.dart
@@ -1,12 +1,15 @@
 import 'dart:async';
+import 'dart:html';
 
 import 'package:angular/angular.dart';
 import 'package:angular_forms/angular_forms.dart';
 import 'package:angular_router/angular_router.dart';
 
+import '../hero_search/hero_search_component.dart';
 import '../hero_component/hero_component.dart';
 import '../../support/server_calls/hero_data_service.dart';
 import '../../route_paths.dart';
+import '../../routes.dart';
 
 import '../../support/data_model/next_generate_hero.dart';
 
@@ -17,9 +20,12 @@ import '../../support/data_model/next_generate_hero.dart';
   directives: [
     coreDirectives,
     formDirectives,
-    HeroComponent
+    HeroComponent,
+    HeroSearchComponent,
   ],
-  providers: [ClassProvider(HeroDataService)],
+  providers: [
+    ClassProvider(HeroDataService),
+  ],
   pipes: [commonPipes],
 )
 class HeroTourListComponent implements OnInit, OnActivate {
@@ -34,6 +40,7 @@ class HeroTourListComponent implements OnInit, OnActivate {
 
   NextGenHero selectedHero;
   List < NextGenHero > heroList;
+  List < NextGenHero > filteredHeroList;
 
   // hero list for all fire heroes
   List < NextGenHero > fireHeroes = [];
@@ -46,28 +53,34 @@ class HeroTourListComponent implements OnInit, OnActivate {
   // hero list for all thunder heroes
   List < NextGenHero > thunderHeroes = [];
 
-  HeroTourListComponent(this._heroDataService, this._router, this._location);
+  HeroTourListComponent(
+    this._heroDataService, 
+    // this._heroDataTransferService,
+    this._router, 
+    this._location
+    );
 
   @override
   void onActivate(_, RouterState current) async {
     this.displayMode = current.toUrl().split("/")[2];
-  }
-
-  void ngOnInit() {
     this._getAllHeroes();
   }
 
+  void ngOnInit() {
+    this.filteredHeroList = new List<NextGenHero>();
+  }
+
   // clikc and show hero details
-  onSelect(NextGenHero hero) {
-    print("Selected hero is: $hero");
-    // selectedHero = hero;
-    print(hero.heroId);
-    // this._router.navigate(path);
+  void onSelect(NextGenHero hero) {
+    this._router.navigate(RoutePaths.hero.toUrl(parameters: {
+      idParam: hero.heroId
+    }));
   }
 
   // get all heroes from server
   Future < void > _getAllHeroes() async {
     this.heroList = await _heroDataService.getAllHeroes();
+    this.filteredHeroList = this.heroList;
     // if have heroes, populate
     if (displayMode == "classified") {
       if (this.heroList.length != 0) {
@@ -84,7 +97,7 @@ class HeroTourListComponent implements OnInit, OnActivate {
   }
 
   // populate hero list according to ability type
-  _populateAbilityHeroes(List < NextGenHero > heroList) {
+  void _populateAbilityHeroes(List < NextGenHero > heroList) {
     if (heroList.length != 0) {
       for (var hero in heroList) {
         switch (hero.abilityType.toUpperCase()) {
@@ -111,39 +124,31 @@ class HeroTourListComponent implements OnInit, OnActivate {
   // delete hero
   Future < void > deleteSelectedHero(NextGenHero hero) async {
     assert(hero.heroId.toString() != "");
-    final result = await this._heroDataService.deleteHero(hero.heroId);
-    if(result == "success") {
-      print("Start to retrieve new heroes list");
-      this._getAllHeroes();
-    } else {
-      print("Delete Failed...");
+    var choice = window.confirm("Are you sure you want to delete this powerful hero?");
+    if (choice) {
+      final result = await this._heroDataService.deleteHero(hero.heroId);
+      if (result == "success") {
+        window.alert("Delete Successful.");
+        this._getAllHeroes();
+      } else {
+        window.alert("Sorry, something is wrong....");
+      }
     }
   }
 
   // delete all heroes
   Future < void > deleteAllHeroes() async {
     print("Delete All Heroes");
-    
+    this._heroDataService.deleteAllHeroes();
   }
 
-  String _heroUrl(int id) => RoutePaths.hero.toUrl(parameters: {
-    idParam: '$id'
-  });
+  // add new hero
+  void addNewHero() {
+    this._router.navigate(RoutePaths.addHero.toUrl());
+  }
 
-  // Future < NavigationResult > gotoDetail() => _router.navigate(_heroUrl(selectedHero.heroId);
-
-  // Future < void > add(String name) async {
-  //   name = name.trim();
-  //   if (name.isEmpty) return null;
-  //   heroList.add(await _heroService.create(name));
-  //   selectedHero = null;
-  // }
-
-  // Future < void > delete(Hero hero) async {
-  //   await _heroService.delete(hero.id);
-  //   heroList.remove(hero);
-  //   if (selectedHero == hero) selectedHero = null;
-  // }
-
-
+  // process search option from hero-search component
+  void getSearchResult(List < NextGenHero > searchResult) {
+    this.filteredHeroList = searchResult;
+  }
 }
